Remove specialized warning for Vue 2
Since Vue 3 no longer has a default export,
ESM doesn't like the import statement as is.
Changing it to `import * as Vue from 'vue'` would work,
but break tree-shaking in projects using this library.

Since Vue 2 is long EOL (but still supported by this library)
and `console.warn` works just as fine for Vue 2
(only with a little less context), removing that specialized
support is fine in a minor version.